refactor(api): move sync to modules/ as controller, model, service
