+++ apps/api/src/modules/sync/service.ts
@@ -1,7 +1,7 @@
 import { db, projectMember, projectRole, revision } from '@repo/db';
 import { and, eq, inArray } from 'drizzle-orm';
 import { toMemberContext, type MemberRole } from '#modules/members/service';
-import { hasPermission, type PermissionResource } from '../shared/permissions';
+import { hasPermission, type PermissionResource } from '#shared/permissions';
 
 // The revision engine's read side. The counters themselves are written by the
 // triggers in migration 0070 — no application code bumps them, so a write moves
