+++ apps/api/src/planner.ts
@@ -37,7 +37,7 @@ import { notificationSettingsRoutes } from './modules/notification-settings';
 import { notificationPreferenceRoutes } from './modules/notification-preferences';
 import { userPreferenceRoutes } from './modules/user-preferences';
 import { telegramRoutes } from './telegram/routes';
-import { syncRoutes } from './sync/routes';
+import { syncRoutes } from './modules/sync';
 
 // The planner API: projects and their columns, issue types, labels, AI agents,
 // custom fields, issues, attachments, saved views, and actions. Mounted on the
